add kind meta to promise object for plan/migration types with missing king
