+++ src/app/debug/duck/utils.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { TreeViewDataItem } from '@patternfly/react-core';
-import { DebugStatusType, IDebugRef, IDebugTreeNode, IDerivedDebugStatusObject } from './types';
+import { IDebugRefWithStatus, IDebugTreeNode } from './types';
 import crawl from 'tree-crawl';
 import TreeActionsDropdown from '../../home/pages/PlanDebugPage/components/TreeActionsDropdown';
 import TreeViewStatusIcon from '../components/TreeViewStatusIcon';
@@ -10,7 +10,7 @@ const uuidv1 = require('uuid/v1');
 const getShallowPropsForNode = (
   rawNode: IDebugTreeNode,
   viewRawDebugObject: (node: IDebugTreeNode) => void,
-  debugRefs?: IDebugRef[]
+  debugRefs?: IDebugRefWithStatus[]
 ): TreeViewDataItem => {
   const matchingDebugRef = debugRefs?.find((ref) => ref?.refName === rawNode?.name);
   return {
@@ -71,134 +71,4 @@ export const convertRawTreeToViewTree = (
       defaultExpanded: true,
     },
   ];
-};
-const calculateCurrentStatus = (hasWarning, hasFailure, hasCompleted, hasRunning) => {
-  let currentStatus;
-  if (hasRunning) {
-    currentStatus = DebugStatusType.Running;
-  } else if (hasFailure) {
-    currentStatus = DebugStatusType.Failure;
-  } else if (hasWarning) {
-    currentStatus = DebugStatusType.Warning;
-  } else if (hasCompleted) {
-    currentStatus = DebugStatusType.Completed;
-  }
-  return currentStatus;
-};
-
-export const getResourceStatus = (debugRef: IDebugRef): IDerivedDebugStatusObject => {
-  switch (debugRef?.kind) {
-    case 'Backup': {
-      const { errors, warnings, phase } = debugRef.status;
-      const hasWarning = warnings?.length > 0 || phase === 'PartiallyFailed';
-      const hasFailure = errors?.length > 0 || phase === 'Failed';
-      const hasCompleted = phase === 'Completed';
-      const hasRunning = phase === 'InProgress';
-      return {
-        hasWarning,
-        hasFailure,
-        hasCompleted,
-        hasRunning,
-        currentStatus: calculateCurrentStatus(hasWarning, hasFailure, hasCompleted, hasRunning),
-      };
-    }
-    case 'Restore': {
-      const { errors, warnings, phase } = debugRef.status;
-      const hasWarning = warnings?.length > 0 || phase === 'PartiallyFailed';
-      const hasFailure = errors?.length > 0 || phase === 'Failed';
-      const hasCompleted = phase === 'Completed';
-      const hasRunning = phase === 'InProgress';
-      return {
-        hasWarning,
-        hasFailure,
-        hasCompleted,
-        hasRunning,
-        currentStatus: calculateCurrentStatus(hasWarning, hasFailure, hasCompleted, hasRunning),
-      };
-    }
-    case 'PodVolumeBackup': {
-      const { phase } = debugRef.status;
-      const hasWarning = phase === 'PartiallyFailed';
-      const hasFailure = phase === 'Failed';
-      const hasCompleted = phase === 'Completed';
-      const hasRunning = phase === 'InProgress';
-      return {
-        hasWarning,
-        hasFailure,
-        hasCompleted,
-        hasRunning,
-        currentStatus: calculateCurrentStatus(hasWarning, hasFailure, hasCompleted, hasRunning),
-      };
-    }
-    case 'PodVolumeRestore': {
-      const { phase } = debugRef.status;
-      const hasWarning = phase === 'PartiallyFailed';
-      const hasFailure = phase === 'Failed';
-      const hasCompleted = phase === 'Completed';
-      const hasRunning = phase === 'InProgress';
-      return {
-        hasWarning,
-        hasFailure,
-        hasCompleted,
-        hasRunning,
-        currentStatus: calculateCurrentStatus(hasWarning, hasFailure, hasCompleted, hasRunning),
-      };
-    }
-    case 'DirectImageMigration': {
-      const { conditions } = debugRef.status;
-      const hasWarning = conditions.some((c) => c.category === ('Critical' || 'Error' || 'Warn'));
-      const hasFailure = conditions.some((c) => c.type === 'Failed');
-      const hasCompleted = conditions.some((c) => c.type === 'Completed');
-      const hasRunning = conditions.some((c) => c.type === 'Running');
-      return {
-        hasWarning,
-        hasFailure,
-        hasCompleted,
-        hasRunning,
-        currentStatus: calculateCurrentStatus(hasWarning, hasFailure, hasCompleted, hasRunning),
-      };
-    }
-    case 'DirectImageStreamMigration': {
-      const { conditions } = debugRef.status;
-      const hasWarning = conditions.some((c) => c.category === ('Critical' || 'Error' || 'Warn'));
-      const hasFailure = conditions.some((c) => c.type === 'Failed');
-      const hasCompleted = conditions.some((c) => c.type === 'Completed');
-      const hasRunning = conditions.some((c) => c.type === 'Running');
-      return {
-        hasWarning,
-        hasFailure,
-        hasCompleted,
-        hasRunning,
-        currentStatus: calculateCurrentStatus(hasWarning, hasFailure, hasCompleted, hasRunning),
-      };
-    }
-    case 'DirectVolumeMigrationProgress': {
-      const { conditions } = debugRef.status;
-      const hasWarning = conditions.some((c) => c.category === ('Critical' || 'Error' || 'Warn'));
-      const hasFailure = conditions.some((c) => c.type === ('InvalidPod' || 'InvalidPodRef'));
-      const hasCompleted = false;
-      const hasRunning = false;
-      return {
-        hasWarning,
-        hasFailure,
-        hasCompleted,
-        hasRunning,
-        currentStatus: calculateCurrentStatus(hasWarning, hasFailure, hasCompleted, hasRunning),
-      };
-    }
-    case 'MigMigration': {
-      const { conditions } = debugRef.status;
-      const hasWarning = conditions.some((c) => c.category === ('Critical' || 'Error' || 'Warn'));
-      const hasFailure = conditions.some((c) => c.type === 'Failed');
-      const hasCompleted = conditions.some((c) => c.type === 'Completed');
-      const hasRunning = conditions.some((c) => c.type === 'Running');
-      return {
-        hasWarning,
-        hasFailure,
-        hasCompleted,
-        hasRunning,
-        currentStatus: calculateCurrentStatus(hasWarning, hasFailure, hasCompleted, hasRunning),
-      };
-    }
-  }
 };
